drm: ocdrm: rework in order to take advantage of drm_kms_simple_helper
Suggested-by: Daniel Vetter <daniel@ffwll.ch>
