+++ drivers/gpu/drm/ocdrm/Makefile
@@ -2,6 +2,6 @@
 # Makefile for the drm device driver.  This driver provides support for the
 # Direct Rendering Infrastructure (DRI) in XFree86 4.1.0 and higher.
 
-ocdrm-y := ocdrm_crtc.o ocdrm_drv.o ocdrm_encoder.o
+ocdrm-y := ocdrm_drv.o
 
 obj-$(CONFIG_DRM_OCDRM) += ocdrm.o
