--- drivers/gpu/drm/ocdrm/ocdrm_encoder.h
@@ -1,48 +1,0 @@
-/*
- * Open cores VGA/LCD 2.0 core DRM driver
- * Copyright (c) 2016 Istituto Italiano di Tecnologia
- * Electronic Design Lab.
- *
- * Author: Andrea Merello <andrea.merello@gmail.com>
- *
- * Based on the following drivers:
- *   - Analog Devices AXI HDMI DRM driver, which is
- *     Copyright 2012 Analog Devices Inc.
- *
- *   - ARC PGU DRM driver.
- *     Copyright (C) 2016 Synopsys, Inc. (www.synopsys.com)
- *
- *   - ARM HDLCD Driver
- *     Copyright (C) 2013-2015 ARM Limited
- *
- *   - Atmel atmel-hlcdc driver, which is
- *     Copyright (C) 2014 Traphandler
- *     Copyright (C) 2014 Free Electrons
- *
- *   - OpenCores VGA/LCD 2.0 core frame buffer driver
- *     Copyright (C) 2013 Stefan Kristiansson, stefan.kristiansson@saunalahti.fi
- *
- *   - R-Car Display Unit DRM driver
- *     Copyright (C) 2013-2015 Renesas Electronics Corporation
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 2 as published by
- * the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along with
- * this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-#ifndef _OCDRM_ENCODER_H_
-#define _OCDRM_ENCODER_H_
-
-#include "ocdrm_drv.h"
-
-int ocdrm_encoder_create(struct ocdrm_priv *priv);
-
-#endif
